+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Rust 1.66.0 is now the earliest officially supported Rust version.
+- IMAP4rev2 is now officially supported.
 - Update OpenSSL bindings and other crate versions to support latest Rust
   version.
 - Added option to redirect standard error to a file so that fatal errors do not